Move cluster functions to it's own namespace
