+++ go/cron/batch.go
@@ -3,6 +3,7 @@ package maint
 import (
 	"context"
 	"database/sql"
+	"github.com/inverse-inc/packetfence/go/log"
 	"time"
 )
 
@@ -17,7 +18,7 @@ func BatchStmt(ctx context.Context, time_limit time.Duration, stmt *sql.Stmt, ar
 
 		rows, err := results.RowsAffected()
 		if err != nil {
-			logError(ctx, "Database error: "+err.Error())
+			log.LogError(ctx, "Database error: "+err.Error())
 			break
 		}
 
@@ -41,7 +42,7 @@ func BatchStmtQueryWithCount(ctx context.Context, time_limit time.Duration, stmt
 		var count int64
 		err := stmt.QueryRow(args...).Scan(&count)
 		if err != nil {
-			logError(ctx, "Database error: "+err.Error())
+			log.LogError(ctx, "Database error: "+err.Error())
 			break
 		}
 
@@ -61,13 +62,13 @@ func BatchStmtQueryWithCount(ctx context.Context, time_limit time.Duration, stmt
 func BatchSql(ctx context.Context, timeout time.Duration, sql string, args ...interface{}) {
 	db, err := getDb()
 	if err != nil {
-		logError(ctx, err.Error())
+		log.LogError(ctx, err.Error())
 		return
 	}
 
 	stmt, err := db.Prepare(sql)
 	if err != nil {
-		logError(ctx, err.Error())
+		log.LogError(ctx, err.Error())
 		return
 	}
 
@@ -78,13 +79,13 @@ func BatchSql(ctx context.Context, timeout time.Duration, sql string, args ...in
 func BatchSqlCount(ctx context.Context, timeout time.Duration, sql string, args ...interface{}) {
 	db, err := getDb()
 	if err != nil {
-		logError(ctx, err.Error())
+		log.LogError(ctx, err.Error())
 		return
 	}
 
 	stmt, err := db.Prepare(sql)
 	if err != nil {
-		logError(ctx, err.Error())
+		log.LogError(ctx, err.Error())
 		return
 	}
